HomeView 파일 구조 리펙토링
